requirements.txt
junitparser==1.4.1
matplotlib==3.3.2
networkx==2.4
numpy==1.19.2
pydot==1.4.1
PyYAML==5.3.1
xlrd==1.2.0
xlwt==1.3.0